refactor: re-arrange stuff
